refactor: 훅 위치 정리
