+++ hooks/nav/useInfinityGoalList.ts
