+++ hooks/nav/useNewGoalInput.ts
